Fix LBFGS options
