+++ config/trajectory_editing_lbfgs.yml
@@ -3,3 +3,5 @@ num_ode_steps: 10
 num_opt_steps: 100
 gradient_checkpoint: False
 optimizer: LBFGS # Adam, SGD, LBFGS, LevenbergMarquardt, ...
+lbfgs:
+  history_size: 10
